Automatically preselect first customer group in price group management dialog
Currently discounts, which are associated to a price group, are lazy-loaded, when the user selects a customer group. Since the save action for price groups simply overwrites the data stored in the database with the data received from the client, this behavior leads to the deletion of associated discounts, if the user only alters the name of a price group without "lazy-loading" the discounts before he clicks the save button. The automatic preselection of the first customer group ensures, that the discounts are always been loaded, when the user triggers the save action.