Fix handling of arrays in the REST API filter
The REST API accepts filter parameters in the URL, which can contain multiple values for one property. The internal handling in the used query builder checks for arrays of filter values, to create "IN" expressions in the resulting SQL query. However, if an array is recognized, the term "IN" will be selected as the operator, but the value array will not be handled correctly. This leads to a syntax error like the following: "Errormesage: [Syntax Error] line 0, col 78: Error: Expected Doctrine\ORM\Query\Lexer::T_OPEN_PARENTHESIS, got ':cleared'".
Adding parentheses, as well as one parameter key for each value in the array and replacing all these keys with their values later fixes this error.

Conflicts:
	engine/Shopware/Components/Model/QueryBuilder.php