Exclude shipping costs from order amount when shipping costs are hidden
When the document is configured to not include the shipping costs as a position, as is the case with delivery notes per default, the order amount should also not include the shipping costs, because otherwise the sum of the positions does not equal the order amount. This does normally not attract attention, but if for example the delivery note template is changed to also show the prices of the positions and the total order amount, the problem becomes obvious.