Include the country id in the address comparison
Before this commit, the address comparison in the backend order details did not include the countries of the billing and shipping addresses. However, since some customers change the country to get around certain country restrictions, it is important to compare the countries, too. This commit adds this check, by comparing the country ids of the two addresses.