Make sure that the article-purchase-unit paragraph has a close tag
If the article has a purchaseunit, but the referenceunit equals the purchaseunit, the article-purchase-unit paragraph has previously not been getting closed. Furthermore, if only the referenceunit is set, but purchaseunit is not, the base price can not be calculated. Because of this, there is no point in displaying the reference price in this case.