Use JSON as logging format
This commit adds a new Monolog formatter that is based on the default `JsonFormatter`. However, the new formatter checks the logged records for an `exception` passed in the context and, if present, adds the exception data like the stack trace to the logged JSON. This allows an easier automated log parsing. In fact, in some places within the Shopware codebase, `exception`s are already passed to the logging context.

This commit does not introduce any breaking changes to Shopware. However, if someone uses automated log parsing, which is based on the old text format, it has to be adopted to parse the JSON instead.