Removed deprecated controller require
Requireing the code and voucher controller with the require property leads to a double load of the module in the backend. Therefore overrides of this controllers are ignored.

The controller is loaded twice because it is loaded first when the module is opened in the backend and a second time when the main controller requires it